requirements.txt
django
django-heroku
gunicorn
requests
pillow
pycrypto
whitenoise
dj-database-url
psycopg2
django-phonenumber-field
